backend/requirements.txt
# ================================================
# BACKEND REQUIREMENTS - VERSIÓN ACTUALIZADA
# ================================================
# Compatible con Azure AI Foundry Agent Service
# Actualizado: Enero 2025

# Core Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Azure AI SDKs - IMPORTANTE: El orden importa
azure-ai-projects  # Usa la última versión disponible
azure-identity==1.19.0
azure-storage-blob==12.23.1
azure-core==1.32.0

# Dependencias adicionales de Azure AI
azure-ai-inference  # Requerido por azure-ai-projects
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0

# Utilidades
python-dotenv==1.0.1
pydantic==2.10.3

# Logging y monitoreo
python-json-logger==3.2.0

# Para desarrollo y testing (opcional)
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0

# NOTA: Si azure.ai.agents.models no funciona, el código debe usar:
# from azure.ai.projects.models import CodeInterpreterTool